Refactor: create renderutils.py and renderables.py
